user-service/pyproject.toml
[tool.poetry]
name = "online-mart"
version = "0.1.0"
description = ""
authors = ["Bilal Raza <success9262@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
aiokafka = "^0.10.0"
sqlmodel = "^0.0.18"
psycopg = "^3.1.19"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
types-python-jose = "^3.3.4.20240106"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
types-passlib = "^1.7.7.20240327"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
